# applications/oxybank/web/openapi/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/stats": {
      "get": {
        "tags": [
          "Annotation Statistics"
        ],
        "summary": "Get overall QA statistics",
        "description": "Get overall statistics data for annotation platform",
        "operationId": "get_overall_stats_api_v1_stats_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data status",
              "title": "Status"
            },
            "description": "Data status"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 4,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Priority",
              "title": "Priority"
            },
            "description": "Priority"
          },
          {
            "name": "data_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data type",
              "title": "Data Type"
            },
            "description": "Data type"
          },
          {
            "name": "caller",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller",
              "title": "Caller"
            },
            "description": "Caller"
          },
          {
            "name": "callee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Callee",
              "title": "Callee"
            },
            "description": "Callee"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data category",
              "title": "Category"
            },
            "description": "Data category"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data tags (comma separated)",
              "title": "Tags"
            },
            "description": "Data tags (comma separated)"
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Creation time start",
              "title": "Start Time"
            },
            "description": "Creation time start"
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Creation time end",
              "title": "End Time"
            },
            "description": "Creation time end"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text search (alias for search_text)",
              "title": "Search"
            },
            "description": "Full-text search (alias for search_text)"
          },
          {
            "name": "search_text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text search",
              "title": "Search Text"
            },
            "description": "Full-text search"
          },
          {
            "name": "trace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Source trace ID",
              "title": "Trace Id"
            },
            "description": "Source trace ID"
          },
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Source group ID",
              "title": "Group Id"
            },
            "description": "Source group ID"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "Sort field"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort direction",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_OverallStatsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/pending-p0": {
      "get": {
        "tags": [
          "Annotation Statistics"
        ],
        "summary": "Get pending P0 data",
        "description": "Get all pending P0 priority data",
        "operationId": "get_pending_p0_stats_api_v1_stats_pending_p0_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Result count limit",
              "default": 100,
              "title": "Limit"
            },
            "description": "Result count limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_PendingP0Response_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stats/by-type": {
      "get": {
        "tags": [
          "Annotation Statistics"
        ],
        "summary": "Get statistics by type",
        "description": "Get statistics by data type (e2e/agent/llm/tool/custom)",
        "operationId": "get_stats_by_type_api_v1_stats_by_type_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data status",
              "title": "Status"
            },
            "description": "Data status"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 4,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Priority",
              "title": "Priority"
            },
            "description": "Priority"
          },
          {
            "name": "data_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data type",
              "title": "Data Type"
            },
            "description": "Data type"
          },
          {
            "name": "caller",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Caller",
              "title": "Caller"
            },
            "description": "Caller"
          },
          {
            "name": "callee",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Callee",
              "title": "Callee"
            },
            "description": "Callee"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data category",
              "title": "Category"
            },
            "description": "Data category"
          },
          {
            "name": "tags",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Data tags (comma separated)",
              "title": "Tags"
            },
            "description": "Data tags (comma separated)"
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Creation time start",
              "title": "Start Time"
            },
            "description": "Creation time start"
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Creation time end",
              "title": "End Time"
            },
            "description": "Creation time end"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text search (alias for search_text)",
              "title": "Search"
            },
            "description": "Full-text search (alias for search_text)"
          },
          {
            "name": "search_text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Full-text search",
              "title": "Search Text"
            },
            "description": "Full-text search"
          },
          {
            "name": "trace_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Source trace ID",
              "title": "Trace Id"
            },
            "description": "Source trace ID"
          },
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Source group ID",
              "title": "Group Id"
            },
            "description": "Source group ID"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page",
              "default": 10,
              "title": "Page Size"
            },
            "description": "Items per page"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Sort field",
              "default": "created_at",
              "title": "Sort By"
            },
            "description": "Sort field"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort direction",
              "default": "desc",
              "title": "Sort Order"
            },
            "description": "Sort direction"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIResponse_TypeStatsResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "APIResponse_OverallStatsResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OverallStatsResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[OverallStatsResponse]"
      },
...
      "APIResponse_PendingP0Response_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PendingP0Response"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[PendingP0Response]"
      },
...
      "APIResponse_TypeStatsResponse_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Response status code",
            "default": 200
          },
          "msg": {
            "type": "string",
            "title": "Msg",
            "description": "Response message",
            "default": "success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TypeStatsResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response data"
          }
        },
        "type": "object",
        "title": "APIResponse[TypeStatsResponse]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OverallStatsResponse": {
        "properties": {
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total data count"
          },
          "pending_count": {
            "type": "integer",
            "title": "Pending Count",
            "description": "Pending count"
          },
          "annotated_count": {
            "type": "integer",
            "title": "Annotated Count",
            "description": "Annotated count"
          },
          "approved_count": {
            "type": "integer",
            "title": "Approved Count",
            "description": "Approved count"
          },
          "rejected_count": {
            "type": "integer",
            "title": "Rejected Count",
            "description": "Rejected count"
          },
          "kb_ingested_count": {
            "type": "integer",
            "title": "Kb Ingested Count",
            "description": "KB ingested count",
            "default": 0
          },
          "kb_pending_count": {
            "type": "integer",
            "title": "Kb Pending Count",
            "description": "KB pending count",
            "default": 0
          },
          "kb_failed_count": {
            "type": "integer",
            "title": "Kb Failed Count",
            "description": "KB failed count",
            "default": 0
          },
          "priority_stats": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Priority Stats",
            "description": "Statistics by priority {priority: count}"
          },
          "type_stats": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Type Stats",
            "description": "Statistics by data type {type: count}"
          }
        },
        "type": "object",
        "required": [
          "total_count",
          "pending_count",
          "annotated_count",
          "approved_count",
          "rejected_count"
        ],
        "title": "OverallStatsResponse",
        "description": "Overall statistics response"
      },
...
      "PendingP0Item": {
        "properties": {
          "data_id": {
            "type": "string",
            "title": "Data Id",
            "description": "Data ID"
          },
          "question": {
            "type": "string",
            "title": "Question",
            "description": "Question"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "description": "Answer"
          },
          "caller": {
            "type": "string",
            "title": "Caller",
            "description": "Caller"
          },
          "callee": {
            "type": "string",
            "title": "Callee",
            "description": "Callee"
          },
          "source_trace_id": {
            "type": "string",
            "title": "Source Trace Id",
            "description": "trace_id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "Creation time"
          }
        },
        "type": "object",
        "required": [
          "data_id",
          "question",
          "answer",
          "caller",
          "callee",
          "source_trace_id",
          "created_at"
        ],
        "title": "PendingP0Item",
        "description": "Pending P0 data item"
      },
      "PendingP0Response": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PendingP0Item"
            },
            "type": "array",
            "title": "Items",
            "description": "Data list"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "PendingP0Response",
        "description": "Pending P0 data response"
      },
...
      "TypeStatsItem": {
        "properties": {
          "data_type": {
            "type": "string",
            "title": "Data Type",
            "description": "Data type"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total count"
          },
          "pending_count": {
            "type": "integer",
            "title": "Pending Count",
            "description": "Pending count"
          },
          "approved_count": {
            "type": "integer",
            "title": "Approved Count",
            "description": "Approved count"
          },
          "rejected_count": {
            "type": "integer",
            "title": "Rejected Count",
            "description": "Rejected count"
          }
        },
        "type": "object",
        "required": [
          "data_type",
          "total_count",
          "pending_count",
          "approved_count",
          "rejected_count"
        ],
        "title": "TypeStatsItem",
        "description": "Type statistics item"
      },
      "TypeStatsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TypeStatsItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Type statistics list"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "TypeStatsResponse",
        "description": "Statistics by type response"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}